requirements.txt
certifi==2018.4.16
chardet==3.0.4
idna==2.6
kafka-python==1.4.2
requests==2.18.4
urllib3==1.22
PyYAML==5.1
influxdb==5.2.0
pylint==2.4.2
graypy==2.1.0
sphinx==2.2.0
sphinx-rtd-theme==0.4.3